Introduce ITypeHandle interface and extract TypeHandle to own file
- Create ITypeHandle interface in CallingConvention/ with all type
  queries needed by ArgIterator and TransitionBlock
- Extract TypeHandle struct from ArgIterator.cs into TypeHandle.cs
- Move GetElemSize helper to static method on ITypeHandle
- Replace all GetRuntimeTypeHandle() escape hatches in ArgIterator (3)
  and TransitionBlock (3) with ITypeHandle method calls
- Port all TypeHandle references in ArgIterator/TransitionBlock to
  ITypeHandle (fields, parameters, locals, return types)
- TypeHandle retains GetRuntimeTypeHandle() for crossgen2-only code
  (WasmLowering) but it is not part of the interface
